Added scraping listing descriptions and modified dumping listings with rapid api
