+++ requirements.txt
@@ -1,0 +1,44 @@
+appnope==0.1.4
+asttokens==2.4.1
+beautifulsoup4==4.12.3
+certifi==2024.2.2
+charset-normalizer==3.3.2
+comm==0.2.2
+debugpy==1.8.1
+decorator==5.1.1
+exceptiongroup==1.2.1
+executing==2.0.1
+idna==3.7
+ipykernel==6.29.4
+ipython==8.25.0
+jedi==0.19.1
+jupyter_client==8.6.2
+jupyter_core==5.7.2
+matplotlib-inline==0.1.7
+nest-asyncio==1.6.0
+numpy==1.26.4
+packaging==24.0
+pandas==2.2.2
+parso==0.8.4
+pexpect==4.9.0
+platformdirs==4.2.2
+prompt_toolkit==3.0.45
+psutil==5.9.8
+ptyprocess==0.7.0
+pure-eval==0.2.2
+Pygments==2.18.0
+python-dateutil==2.9.0.post0
+python-dotenv==1.0.1
+pytz==2024.1
+pyzmq==26.0.3
+requests==2.32.3
+six==1.16.0
+soupsieve==2.5
+stack-data==0.6.3
+tornado==6.4
+tqdm==4.66.4
+traitlets==5.14.3
+typing_extensions==4.12.1
+tzdata==2024.1
+urllib3==2.2.1
+wcwidth==0.2.13
